fixes acciente/oacc-db#3 - changes case of table identifiers in SQL statements to match the database setup scripts
MySQL on Linux is case-sensitive in regards to table names (because tables are stored as files on the file system). The
SQL statements didn't match in case for one of the tables as it was defined in the DDL setup scripts in oacc-db. This
would cause an error on case-sensitive database environments. The fix was to match the table names exactly.